Change how blocked numbers are removed from the Blacklist DB
Blocked numbers were being removed by their DB._id, change the removal
to be by the number itself. This change is needed for the messaging App
so that when a blocked number is deleted, as opposed to being updated,
in the framework DB, the deleted number can be determined by the
messaging App by looking at the last segment of the Uri that is
passed to the messaging App framework blacklist DB content observer.
Prior to this change, the last segment of the Uri contained the DB._id
of the deleted number in the framework blacklist DB

Change-Id: I03740f38a81f0b4ef7f8965d57e19f5328c68e5c
Ticket: CYNGNOS-2112